+++ tests/Unit/Status/JobInfoTest.php
@@ -3,6 +3,7 @@
 namespace Tests\Orisai\Scheduler\Unit\Status;
 
 use DateTimeImmutable;
+use DateTimeZone;
 use Generator;
 use Orisai\Scheduler\Status\JobInfo;
 use PHPUnit\Framework\TestCase;
@@ -24,14 +25,16 @@ final class JobInfoTest extends TestCase
 		int $repeatAfterSeconds,
 		int $runSecond,
 		DateTimeImmutable $start,
+		?DateTimeZone $timeZone,
 		string $extendedExpression
 	): void
 	{
-		$info = new JobInfo($id, $name, $expression, $repeatAfterSeconds, $runSecond, $start);
+		$info = new JobInfo($id, $name, $expression, $repeatAfterSeconds, $runSecond, $start, $timeZone);
 		self::assertSame($id, $info->getId());
 		self::assertSame($name, $info->getName());
 		self::assertSame($expression, $info->getExpression());
 		self::assertSame($repeatAfterSeconds, $info->getRepeatAfterSeconds());
+		self::assertSame($timeZone, $info->getTimeZone());
 		self::assertSame($extendedExpression, $info->getExtendedExpression());
 		self::assertSame($runSecond, $info->getRunSecond());
 		self::assertSame($start, $info->getStart());
@@ -58,6 +61,7 @@ final class JobInfoTest extends TestCase
 			0,
 			0,
 			new DateTimeImmutable(),
+			null,
 			'* * * * *',
 		];
 
@@ -68,7 +72,30 @@ final class JobInfoTest extends TestCase
 			10,
 			15,
 			new DateTimeImmutable(),
+			null,
 			'* * * * */5 / 10',
+		];
+
+		yield [
+			'dunno',
+			'still dunno',
+			'* * 6 9 *',
+			0,
+			2,
+			new DateTimeImmutable(),
+			new DateTimeZone('Europe/Prague'),
+			'* * 6 9 * (Europe/Prague)',
+		];
+
+		yield [
+			'whatever',
+			'whatever but pale blue',
+			'* 1 9 8 4',
+			30,
+			2,
+			new DateTimeImmutable(),
+			new DateTimeZone('UTC'),
+			'* 1 9 8 4 / 30 (UTC)',
 		];
 	}